requirements.txt
discord.py==2.5.1
Jinja2==3.1.2
html2image==0.3.2
